# web/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BeautyMS API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/coupon/list": {
            "get": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Получение списка купонов",
                "operationId": "8458d4d381c3e83c6eabf0b6c37aba14",
                "responses": {
                    "200": {
                        "description": "Список купонов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "category": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "color": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "format": "uri"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "amount": {
                                                "type": "number"
                                            },
                                            "price": {
                                                "type": "number"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string",
                                                "format": "uri"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/coupon/order": {
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Заказ купона",
                "operationId": "0496acf207d5bce66cf65eed743170c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coupon_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "URL для оплаты",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "format": "uri"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка запроса"
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/coupon/order-list": {
            "get": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Получение списка заказов купонов",
                "operationId": "e4f07684189f23e9a5039d481222b9fa",
                "responses": {
                    "200": {
                        "description": "Список заказов купонов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "coupon": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer"
                                                            },
                                                            "name": {
                                                                "type": "string"
                                                            },
                                                            "color": {
                                                                "type": "string"
                                                            },
                                                            "image": {
                                                                "type": "string",
                                                                "format": "uri"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "amount": {
                                                        "type": "number"
                                                    },
                                                    "price": {
                                                        "type": "number"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "format": "uri"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "status": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}